[#6] Refactor: 이슈 템플릿, 디렉토리 구조 변경
