+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,6 @@
+## 📝 작업 내용
+> <작업한 내용을 간략히 설명해주세요 (스크린샷 첨부 가능)>
+
+
+## 🔍 테스트 방법
+> 1. <테스트 방법을 상세히 적어주세요 (스크린샷 첨부 가능)>
