+++ .gitignore
@@ -35,3 +35,6 @@ out/
 
 ### VS Code ###
 .vscode/
+
+### 추가
+/src/main/resources/*.yml
