+++ src/main/java/com/parentsgowork/server/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.parentsgowork.server;
+package com.parentsgowork.server.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
