+++ src/main/java/com/parentsgowork/server/domain/User.java
@@ -1,4 +1,4 @@
-package com.parentsgowork.server.project.domain;
+package com.parentsgowork.server.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
