+++ src/main/java/com/parentsgowork/server/web/controller/TestController.java
@@ -1,4 +1,4 @@
-package com.parentsgowork.server.project.web.controller;
+package com.parentsgowork.server.web.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
